Change "check out, review, and merge locally" URL to SSH by default.
The default URL to show on the project screen is the SSH URL, so it
makes sense for it to be the default here too. Plus, since most
people (hopefully) use SSH keys to avoid typing credentials repeatedly,
changing this default means these commands can be pasted into a
terminal more easily since no password prompt will get in the way.